Character animation blueprint fixed
Locomotion and crouch blend spaces fixed in character animation blueprint

+++ Source/ThirdPersonShooter/Private/Actors/PickupWeapon.cpp
@@ -181,7 +181,7 @@ void APickupWeapon::SpawnProjectile()
 {
 	if (CurrentProjectile)
 	{
-		for(int i = 0; i < CurrentProjectile->NumberOfPellets; ++i)
+		for(uint8 i = 0; i < CurrentProjectile->NumberOfPellets; ++i)
 		{
 			FVector Location;
 			FRotator Rotation;
